+++ R/piping-model.R
@@ -980,7 +980,7 @@ rxSetCovariateNamesForPiping <- function(covariates=NULL) {
         .minfo(paste0("add between subject variability {.code ", var, "} and set estimate to {.number ", value, "}"))
       }
     }
-    assign("iniDf", rbind(.iniDf, .extra), envir=rxui)
+    assign("iniDf", rbind(.iniDf, .iniDfMatchColumns(.extra, .iniDf)), envir=rxui)
   } else {
     if (is.na(promote)) {
     } else if (!promote) {
@@ -1015,7 +1015,7 @@ rxSetCovariateNamesForPiping <- function(covariates=NULL) {
       } else if (promote) {
         .minfo(paste0("promote {.code ", var, "} to population parameter with initial estimate {.number ", value, "}"))
         # need to reassess model for mu2 enhancement
-        assign("iniDf", rbind(.iniDf, .extra), envir=rxui)
+        assign("iniDf", rbind(.iniDf, .iniDfMatchColumns(.extra, .iniDf)), envir=rxui)
         rxui2 <- rxui
         model(rxui2) <- rxui$lstExpr
         rxui2 <- rxUiDecompress(rxui2)
@@ -1027,7 +1027,7 @@ rxSetCovariateNamesForPiping <- function(covariates=NULL) {
         .minfo(paste0("add population parameter {.code ", var, "} and set estimate to {.number ", value, "}"))
       }
     }
-    assign("iniDf", rbind(.iniDf, .extra), envir=rxui)
+    assign("iniDf", rbind(.iniDf, .iniDfMatchColumns(.extra, .iniDf)), envir=rxui)
   }
   invisible()
 }